Don't wildcard allow-origin by default. (#144)
* Don't wildcard allow-origin by default.

This could be considered a vulnerability in how Plumber behaved previously. This, by default, allows cross-origin GET, HEAD, and POST requests using the standard headers from any origin. We've decided that this is something you should opt-in to, so we're removing this as the default.

* Support OPTIONS requests

Makes it feasible for you to build your own CORS handler.

* Note NEWS changes.