add sourceUTF8 and helper functions from shiny
This stems from wanting to read utf8 files on a non utf8 locale.  On windows, the locale is set to the system default which typically does not support utf8 out of the box.

sourceUTF8 functions are adopted form the rstudio/shiny repo in an effort to use code that has already be validated to work.

The ending finding is that `source(file, encoding="UTF-8")` does not work as expected and instead should use `eval(parse(file, encoding = "UTF-8"))`