Add optional `sameSite` flag to cookies
Allows the SameSite attribute to be set on cookies.
This is especially important for Cross-Origin requests,
as in the near future browsers will automatically default
to the "SameSite=Lax" policy for cookies without an
explicit policy set. This will prevent cookies from being
sent on Cross-Origin requests unless the policy is set to
"None".

See:
https://www.chromestatus.com/feature/5088147346030592
https://www.chromestatus.com/feature/5633521622188032